pipeline/requirements.txt
# Global Flow Radar - Pythonパイプライン依存パッケージ
yfinance>=0.2.31
fredapi>=0.5.1
pandas>=2.1.0
numpy>=1.25.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
redis>=5.0.0
python-dotenv>=1.0.0
requests>=2.31.0
schedule>=1.2.0
